sdcc (2.5.99+2.6.0rc1-1) experimental; urgency=low

  * New upstream release candidate:
    - bool datatype support for z80 target (closes: bug#325846)
    - support for code segments in memory for pic target (closes: bug#359675)
    - support for almost all PIC (closes: bug#361229)
    - support for tables for pic target (closes: bug#304320)
    - support for main() to be in any object when linking (closes: bug#335711)
  * Bumped Standards-Version to 3.7.2 (no changes).

+++ device/lib/modff.c
@@ -22,6 +22,6 @@
 
 float modff(float x, float * y)
 {
-    *y=((int)x);
+    *y=(long)x;
     return (x-*y);
 }
